pyproject.toml
[project]
name = "gamedatagen"
version = "0.1.0"
description = "AI-powered game content generation framework - Generate worlds, NPCs, quests, items with LLMs"
authors = [
    {name = "GameDataGen Contributors"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["game", "mmorpg", "ai", "content-generation", "llm", "procedural-generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Games/Entertainment",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",           # Beautiful CLI output
    "pydantic>=2.0.0",        # Schema validation (replaces Zod)
    "litellm>=1.0.0",         # Multi-model LLM support
    "openai>=1.0.0",          # OpenAI API (for DALL-E, GPT-4o-mini vision)
    "fastapi>=0.110.0",       # API server
    "uvicorn>=0.27.0",        # ASGI server
    "httpx>=0.26.0",          # HTTP client
    "pyyaml>=6.0",            # YAML config
    "python-multipart>=0.0.6", # File uploads for API
    "aiosqlite>=0.19.0",      # Async SQLite
    "numpy>=1.26.0",          # For spatial system
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
gamedatagen = "gamedatagen.cli:main"

[project.urls]
Homepage = "https://github.com/ZachHandley/GameDataGen"
Repository = "https://github.com/ZachHandley/GameDataGen"
Documentation = "https://github.com/ZachHandley/GameDataGen#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gamedatagen"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.black]
line-length = 100
target-version = ["py311", "py312"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
